rename: 컴포넌트 파일명 index.tsx로 변경
