+++ src/components/IssueDetail/index.tsx
@@ -3,7 +3,7 @@ import { useParams } from 'react-router-dom';
 import { getIssueDetail } from '../../api/Api';
 import { IssueType } from '../../utils/types/Issue.interface';
 import ReactMarkdown from 'react-markdown';
-import IssueItem from '../IssueItem/IssueItem';
+import IssueItem from '../IssueItem';
 import LoadingSpinner from '../common/LoadingSpinner';
 import styles from './IssueDetail.module.scss';
