+++ src/pages/IssueDetailPage/index.tsx
@@ -1,4 +1,4 @@
-import IssueDetail from '../../components/IssueDetail/IssueDetail';
+import IssueDetail from '../../components/IssueDetail';
 
 export default function IssueDetailPage() {
 	return <IssueDetail />;
