+++ src/pages/IssueListPage/index.tsx
@@ -1,4 +1,4 @@
-import IssueList from '../../components/IssueList/IssueLIst';
+import IssueList from '../../components/IssueList';
 export default function IssueListPage() {
 	return <IssueList />;
 }
